// src/app/models/personal_details.ts
export interface PersonalDetails {
    firstName: string;
    lastName: string;
    middleName: string;
    employer: string;
    jobTitle: string;
    address: string;
    mobileNumber: string;
    selectedGender: string | null;
    selectedCivilStatus: string | null;
    selectedCountry: string | null;
    selectedCity: string | null;
}